move data writer to separate file
